+++ examples/transport/buttons.php
@@ -9,26 +9,32 @@ return [
                 'bicycle' => [
                     'class' => 'fas fa-bicycle',
                     'value' => 'bicycle',
+                    'title' => 'Bicycle',
                 ],
                 'bus' => [
                     'class' => 'fas fa-bus',
                     'value' => 'bus',
+                    'title' => 'Bus',
                 ],
                 'car' => [
                     'class' => 'fas fa-car',
                     'value' => 'car',
+                    'title' => 'Car',
                 ],
                 'plane' => [
                     'class' => 'fas fa-plane',
                     'value' => 'plane',
+                    'title' => 'Plane',
                 ],
                 'ship' => [
                     'class' => 'fas fa-ship',
                     'value' => 'ship',
+                    'title' => 'Ship',
                 ],
                 'subway' => [
                     'class' => 'fas fa-subway',
                     'value' => 'subway',
+                    'title' => 'Subway',
                 ],
             ],
         ],